watchdog/pyproject.toml
[tool.poetry]
name = "orcha-watchdog"
version = "0.1.3"
description = "Orcha pluggable for SystemD Watchdog integration"
authors = ["Javier Alonso <jalonso@teldat.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7.2"
orcha = "0.3.0-rc2"
systemd-python = "^234"

[tool.poetry.dev-dependencies]
black = "^22.10.0"
flake8 = "^5.0.4"
pylint = "^2.15.5"
flake8-pylint = "^0.1.3"
isort = "^5.10.1"
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
